Add all test files to ya.make
commit 9c2d66210d4f1d5dd6b4b7df595ce84a241ae827
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Wed Jan 29 14:22:09 2020 +0300

    Fix some scheduler tests

commit e816ea228092d59a694f65f83149a51a71807f21
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Wed Jan 29 13:52:58 2020 +0300

    Finish rebase

commit c5d6ad0083ab3aec7764de65c608a04a7ee07be3
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Tue Jan 28 14:54:27 2020 +0300

    Move yson performance tests to separate ya.make

commit 48a38f8b87792026e4959f73d315aa8150446a2c
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Tue Jan 28 12:59:20 2020 +0300

    Mark tests as using suid tool

commit 8cf84c304b444a4004e260660a43a26a55ed7c9c
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Tue Jan 28 12:57:04 2020 +0300

    Fix p2p test on distbuild

commit 6d97a4e54c24f7a9f96aab2b5ccb8827e5635844
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Sun Jan 26 22:23:43 2020 +0300

    Flag test as using REQUIRE_SUID_TOOL

commit 9b0a6bae7969be9d8b70f868ec347e2bcfc23339
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Sun Jan 26 22:16:12 2020 +0300

    Fix some scheduler tests.

commit cb399115f10e788916f0c7d5f022175d4cc323cd
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Sun Jan 26 21:46:44 2020 +0300

    Fix YT-12096

commit 7e12abbdc4f7cd197d1421b149be86c1cdb45cab
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Sun Jan 26 16:17:45 2020 +0300

    Splitting tests.

    - Add missing import ot dynamic_tables_state_transitions
    - Split test_scheduler_common in to two files.

commit e1aa605adef2911832933ef962cbb6ce492d34d5
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Sun Jan 26 03:07:21 2020 +0300

    Split test_scheduler_common in half.

commit 542c2a23d9c67223c78d505d5d754d4d1d5b4c92
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Sun Jan 26 03:05:20 2020 +0300

    Use arcadia gdb

commit 05ca728e47eb6c8e64d9d18ab06f7e5e49cba060
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Sun Jan 26 03:03:58 2020 +0300

    Skip tests USE_PORTO_FOR_SERVERS tests on distbuild.

commit 98289b472dfec55c9ef9eb18362198aa14a66b50
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Sun Jan 26 03:03:12 2020 +0300

    Move state transitions test to another file.

commit c654ebd0c55eead1c988edbe171fd0593fe7ed39
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Sun Jan 26 01:46:17 2020 +0300

    Add rest of the test files to ya.make

commit 78ba15c1207cf79233506e6db99ca3ba07c23fa8
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Sun Jan 26 01:44:53 2020 +0300

    Merge disk_usage and disk_quota into a single file

commit efcbcacc26f1975b3ce92cb59ae994350a89800c
Author: Fedor Korotkiy <prime@yandex-team.ru>
Date:   Sat Jan 25 21:58:49 2020 +0300

    Moving tests to ya.make

    - Enable scheduler simulator tests.
    - Remove test_grpc.py completely. Not sure why this test exists.
    - Disable yson performance tests.
    - Move compression test cases to separate ya.make.
    - Enable test_yson_format.py.

+++ yt/server/node/data_node/peer_block_distributor.cpp
@@ -133,7 +133,7 @@ bool TPeerBlockDistributor::ShouldDistributeBlocks()
     i64 totalRequestedBlockSize = TotalRequestedBlockSize_;
 
     bool shouldDistributeBlocks =
-        outTraffic > Config_->OutTrafficActivationThreshold * Config_->IterationPeriod.MilliSeconds() / 1000.0 ||
+        outTraffic > static_cast<double>(Config_->OutTrafficActivationThreshold) * Config_->IterationPeriod.MilliSeconds() / 1000.0 ||
         totalOutQueueSize > Config_->OutQueueSizeActivationThreshold ||
         totalRequestedBlockSize > Config_->TotalRequestedBlockSizeActivationThreshold * Config_->IterationPeriod.MilliSeconds()  / 1000.0;
